+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Next]
+
+**Changed**
+- Allow `Holotheca` to use identifiers that aren't Integer or String; now any value is supported, with baked-in conversion added for Symbol.
+- Make `Holotheca` more resilient to different `.identifier_format`s, should now support pretty much any case, including re-setting it to `nil`.
+
+**Fixed**
+- `Holotheca#path`, `#root` and `#depth` now all work correctly even if holarchy is changed after instantiation, checking the actual parents, not classes.
+
 ## [v0.1.0] — 2025-04-07
 
 First really useful version.